include: Move utils.h out of the public headers directory
This makes it more explicit that the microdns/ directory is for public
headers only
